requirements.txt
Flask==0.10.1
Jinja2==2.8
Werkzeug==0.10.4
itsdangerous==0.24
PyMongo==3.0.3
wsgiref==0.1.2
PyGithub==1.25.2
Flask-RESTful==0.3.4
Flask-Script==2.0.5
